requirements.txt
esdk-obs-python==3.21.4
pandas==2.0.3
loguru==0.6.0